Moved method to script_functions.
